refactor: tweaked component test structure
